# server/openapi.yaml
...
openapi: 3.0.3
info:
  title: GWP API
  version: 0.1.0
  description: >
    GWP — Canvas Online as a self-hosted app. Phase 0 ships first-run
    setup, session auth and health probes; Phase 1 adds the organization
    surface (company profile, departments, teams — spec §3/§4). Canvas,
    coaching and AI endpoints arrive in later phases.
  contact:
    name: GWP operators
servers:
  - url: /
    description: Same origin as the served web app.
...
paths:
...
  /api/v1/canvases:
    post:
      tags: [canvas]
      summary: Create a canvas and its first draft
      description: >
        Subject policy on ownerUserId: a member may create only for itself,
        a manager for users in its current reporting subtree, the owner for
        anyone — admin for itself only (canvas content is subject data, not
        org metadata). ownerUserId must be an ACTIVE user of the company.
        `name` is the canvas record's display name and is stamped into
        body.meta.title. `body` is the canonical canvas document
        (docs/contracts/canvas-payload.md), validated draft-mode: schema or
        referential errors are 400 with issue paths, blanks are fine.
        assignee_user_id fields inside the body must resolve to
        same-company users or the write is rejected — they are content,
        never an access grant. Creates the canvas and the shared draft
        (revision 1) atomically and audits canvas.create +
        canvas.draft.create (metadata only).
      operationId: createCanvas
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasCreateRequest"
      responses:
        "201":
          description: Canvas and first draft created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasCreateResponse"
        "400":
          description: >
            Body failed validation — strict envelope, a schema-invalid
            canvas body (issue paths in details.fields), or an
            assignee_user_id that is not a user of this company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            No subject access to ownerUserId, or no such user in this
            company — the denial is uniform.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The target owner is not an active user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [canvas]
      summary: List visible canvases (keyset page)
      description: >
        The canvases whose owner sits in the caller's subject scope —
        self for members, the whole company for the owner, the current
        reporting subtree for managers, own canvases only for admin.
        Ordered newest-first over (created_at, id); limit defaults to 25,
        max 100; cursor is the opaque token from nextCursor. An empty
        scope is a valid empty page.
      operationId: listCanvases
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - name: cursor
          in: query
          required: false
          description: Opaque keyset cursor from the previous page's nextCursor.
          schema:
            type: string
      responses:
        "200":
          description: One page of visible canvases.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasPage"
        "400":
          description: Invalid limit or cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}:
    get:
      tags: [canvas]
      summary: Read one canvas (detail + draft + current version summary)
      description: >
        Subject policy on the canvas owner — self/owner/subtree only;
        admin sees only its own canvases. Missing, foreign and denied ids
        all return the same 404. The draft carries its working body; the
        current published version is summarized (its body comes from the
        versions endpoint).
      operationId: getCanvas
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      responses:
        "200":
          description: The canvas detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canvas"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such canvas, or no subject access to its owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [canvas]
      summary: Rename the canvas record
      description: >
        Changes the record's display name only (lists, detail, dashboard).
        Same write gate as every canvas mutation — subject access on the
        canvas owner (self, owner role, manager over the current subtree);
        admin only for its own canvases. Missing, foreign and denied ids
        are the uniform 404, and that check precedes body validation. The
        shared draft (whose meta.title is user-edited content) is NOT
        touched and its revision does not move. Audited as `canvas.rename`
        with the field name only — never the new name.
      operationId: renameCanvas
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasRenameRequest"
      responses:
        "200":
          description: The canvas detail with the new name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canvas"
        "400":
          description: Empty/over-long name or an unknown field (strict body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such canvas, or no subject access to its owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: CANVAS_ARCHIVED — archived canvases are read-only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/draft:
    post:
      tags: [canvas]
      summary: Open the canvas's shared draft
      description: >
        Creates the single shared draft for this canvas — one draft per
        canvas, ever present at most once. The draft body is copied from
        the current published version, or a blank canonical canvas when
        nothing was published yet; revision starts at 1 and
        baseVersionId links the published base. A canvas that already has
        a live draft returns 409 DRAFT_EXISTS — in-progress work is never
        overwritten. Archived canvases return 409 CANVAS_ARCHIVED. Subject
        policy applies exactly as for reads.
      operationId: createCanvasDraft
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      responses:
        "201":
          description: Draft created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasDraft"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such canvas, or no subject access to its owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A draft already exists, or the canvas is archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [canvas]
      summary: Save the shared draft (optimistic concurrency)
      description: >
        Conflict-safe draft write (task 2.4): the client asserts the
        `expectedRevision` it read and the `baseVersionId` the draft
        branched from — the save applies only while both still hold, via
        an UPDATE … WHERE revision CAS under the company write lock. A
        stale revision or base returns 409 DRAFT_CONFLICT and writes
        nothing; a canvas with no draft is the uniform 404. The body is
        re-validated draft-mode (400 with issue paths) and
        assignee_user_id references re-checked exactly like on create.
        Audits canvas.draft.save, metadata only.
      operationId: saveCanvasDraft
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasDraftSaveRequest"
      responses:
        "200":
          description: The saved draft — revision bumped by one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasDraft"
        "400":
          description: >
            Strict-envelope or draft-mode validation failure — issue
            paths in details.fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            No such canvas, no draft to save, or no subject access to
            the canvas owner — the denial is uniform.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            DRAFT_CONFLICT — expectedRevision or baseVersionId is stale;
            or CANVAS_ARCHIVED on an archived canvas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/publish:
    post:
      tags: [canvas]
      summary: Publish the live draft as an immutable version
      description: >
        Turns the shared draft into the next immutable canvas_version
        (task 2.4): version_no = max(version_no)+1 taken under the canvas
        row lock, the draft is consumed, canvas.current_version_id points
        at the new snapshot, and the whole thing commits in one
        transaction with its metadata-only audit row. The body must pass
        publish-mode completeness — errors are 400 with issue paths.
        `idempotencyKey` binds the write to a (scope, key) write_receipt:
        a retry with the same key replays the stored versionId/versionNo
        without writing again, and the same key carrying a different
        request is 409. The receipt is consulted AFTER the permission
        check, so a revoked actor can never replay. `expectedRevision` is
        the draft CAS — a stale value is 409 DRAFT_CONFLICT.
      operationId: publishCanvas
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasPublishRequest"
      responses:
        "200":
          description: The published version (fresh or replayed).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasPublishResponse"
        "400":
          description: >
            Strict-envelope failure or publish-mode completeness errors —
            issue paths in details.fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            No such canvas, no live draft, or no subject access to the
            canvas owner — the denial is uniform.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            IDEMPOTENCY_CONFLICT (key reused with a different payload),
            DRAFT_CONFLICT (stale expectedRevision or base), or
            CANVAS_ARCHIVED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/versions/{versionId}/restore:
    post:
      tags: [canvas]
      summary: Restore a published version into the shared draft
      description: >
        Copies the version's body into the canvas's draft (task 2.4) —
        history is never mutated; restore writes only canvas_draft. With
        no live draft a new one is created (revision 1) based on the
        CURRENT head, content starting at the restored body. With a live
        draft the request must carry its `expectedRevision` AND
        `confirm: true` — restore never silently discards in-progress
        work and answers 409 DRAFT_CONFLICT otherwise. Audits
        canvas.restore, metadata only.
      operationId: restoreCanvasVersion
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
        - $ref: "#/components/parameters/VersionId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasRestoreRequest"
      responses:
        "200":
          description: The draft holding the restored body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasDraft"
        "400":
          description: Strict-envelope validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            No such canvas/version pair, or no subject access to the
            canvas owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            DRAFT_CONFLICT — a live draft exists and expectedRevision +
            confirm:true were not both supplied; or CANVAS_ARCHIVED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/archive:
    post:
      tags: [canvas]
      summary: Archive the canvas (write-off flag, never a delete)
      description: >
        Flips status to archived with archived_at/archived_by (task 2.4,
        spec §9). Reads keep working; every write path — save, publish,
        draft open, restore, transfer — answers 409 CANVAS_ARCHIVED.
        Archiving an already-archived canvas is the same 409. Audits
        canvas.archive, metadata only.
      operationId: archiveCanvas
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      responses:
        "200":
          description: The canvas detail after archiving.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canvas"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such canvas, or no subject access to its owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The canvas is already archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/transfer:
    post:
      tags: [canvas]
      summary: Transfer canvas ownership (owner role only)
      description: >
        Reassigns the canvas owner (task 2.4). Restricted by ROLE, not by
        subject scoping: the caller must hold the owner role — member,
        manager and admin are all denied 403 even where subject access
        would pass. The target must be an ACTIVE user of the same
        company — foreign/unknown ids are the uniform 404, an inactive
        one is 409 USER_NOT_ACTIVE. Access follows the new owner
        immediately. Audits canvas.transfer, metadata only.
      operationId: transferCanvasOwner
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasTransferRequest"
      responses:
        "200":
          description: The canvas detail under the new owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canvas"
        "400":
          description: Strict-envelope validation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: The caller does not hold the owner role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            No such canvas, no such target user, or no subject access to
            the canvas owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The target user is inactive or the canvas is archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/versions:
    get:
      tags: [canvas]
      summary: List published versions of a canvas (history)
      description: >
        Every published version of the canvas, newest first — summaries
        only (no bodies; each snapshot is fetched through
        GET .../versions/{versionId}). Versions are bounded per canvas —
        publish is a deliberate manual action — so the full ordered list
        is returned without pagination. Subject policy on the canvas
        owner applies to history exactly as to drafts: missing, foreign
        and denied ids all return the same 404.
      operationId: listCanvasVersions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      responses:
        "200":
          description: The version history, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CanvasVersionSummary"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No such canvas, or no subject access to its owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/versions/{versionId}:
    get:
      tags: [canvas]
      summary: Read one published version snapshot
      description: >
        Immutable snapshot read — there is no update or delete route for
        versions (the runtime DB role cannot mutate canvas_version at
        all). Subject access on the canvas owner applies to history the
        same as to drafts (spec §4); a version id paired with another
        canvas's id is a uniform 404.
      operationId: getCanvasVersion
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
        - $ref: "#/components/parameters/VersionId"
      responses:
        "200":
          description: The published snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasVersion"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            No such canvas/version pair, or no subject access to the
            canvas owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/canvases/{id}/versions/{versionId}/export:
    get:
      tags: [canvas]
      summary: Export one published version (JSON or Markdown)
      description: >
        Audited export boundary (task 2.7, spec §5.4): reads the immutable
        published snapshot and returns it either verbatim as the canonical
        JSON body or rendered to Markdown with explicit loss warnings —
        extension fields Markdown cannot carry are named in `warnings`, so
        callers can fall back to the lossless JSON. Same subject policy as
        a version read: denied, foreign, missing and mismatched canvas/
        version pairs all return a uniform 404. Every export is recorded in
        the audit trail as metadata only ({mode, version}) — exported
        content never enters the audit log.
      operationId: exportCanvasVersion
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CanvasId"
        - $ref: "#/components/parameters/VersionId"
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, markdown]
            default: json
      responses:
        "200":
          description: The exported snapshot plus loss warnings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasVersionExport"
        "400":
          description: format is not json|markdown.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            No such canvas/version pair, or no subject access to the
            canvas owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    LimitParam:
      name: limit
      in: query
      required: false
      description: Page size — default 25, maximum 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    CanvasId:
      name: id
      in: path
      required: true
      description: Canvas id inside the caller's company.
      schema:
        type: string
        format: uuid
    VersionId:
      name: versionId
      in: path
      required: true
      description: Published version id of the canvas.
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Stable machine code (e.g. INVALID_SESSION).
        message:
          type: string
          description: Human-readable, Vietnamese UI copy.
        details:
          description: Optional structured detail (field names, pending list).
        request_id:
          type: string
          format: uuid
          description: Echo of the X-Request-Id response header.
...
    CanvasCreateRequest:
      type: object
      required: [ownerUserId, name, body]
      additionalProperties: false
      properties:
        ownerUserId:
          type: string
          format: uuid
          description: >
            The subject the canvas belongs to — must be an active user of
            the caller's company and inside the caller's subject scope.
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Display name; stamped into body.meta.title.
        body:
          $ref: "#/components/schemas/CanvasBody"
    CanvasCreateResponse:
      type: object
      required: [id, draft]
      properties:
        id:
          type: string
          format: uuid
        draft:
          type: object
          required: [id, revision]
          properties:
            id:
              type: string
              format: uuid
            revision:
              type: integer
...
    CanvasPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CanvasListItem"
        nextCursor:
          type: string
          nullable: true
          description: Pass as ?cursor= for the next page; null at the end.
...